feat(onboardingSurvey): add backgroundGenerator
